requirements.txt
opencv-python>=4.7.0.72
numpy>=1.24.2
Pillow>=9.4.0
torchvision>=0.14.1